move base node store slice to /store/baseNode
