+++ applications/launchpad_v2/src/containers/BaseNodeContainer/types.ts
@@ -1,16 +1,10 @@
-export type Network = 'mainnet' | 'testnet'
+import type { Network } from '../../store/baseNode/types'
 
 export interface BaseNodeProps {
   startNode: () => void
   stopNode: () => void
   tariNetwork: Network
   setTariNetwork: (selected: Network) => void
-  running: boolean
-  pending: boolean
-}
-
-export type BaseNodeState = {
-  network: Network
   running: boolean
   pending: boolean
 }
