+++ applications/launchpad_v2/src/store/baseNode/types.ts
@@ -1,0 +1,7 @@
+export type Network = 'mainnet' | 'testnet'
+
+export type BaseNodeState = {
+  network: Network
+  running: boolean
+  pending: boolean
+}
