+++ applications/launchpad_v2/src/store/index.ts
@@ -1,7 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit'
 
-import baseNodeReducer from '../containers/BaseNodeContainer/slice'
-
+import baseNodeReducer from './baseNode'
 import appReducer from './app'
 
 export const store = configureStore({
